move updateToast to subfolder
