+++ frontend/src/components/Toast/UpdateToast/updateToastStyle.ts
